confirmed latest klayout and python work

+++ PY_klayout.py
@@ -1,5 +1,4 @@
-import klayout.db as db #require 0.26.12
-
+      
 # def format_execute_lsf():
         # #export settings to lsf
         # mysettings = #
@@ -18,47 +17,61 @@ import klayout.db as db #require 0.26.12
         # write("export_settings.lsf",mysettings,"overwrite");
         # export_settings; #lsf file to run.
 
-def klayout_formatfile(input="output.gds"):
-    #Due to a format update in Klayout 0.27+, newer version of Klayout cannot read the file format exported
-    #by the lumerical encrypted library. The work around is to load the file in Klayout26 and resave it.
-    #I believe the issue is because the Lumerical encrpyted library is saving multiple topcells
-    #This can be seen from the errors when loading them directly into Klayout 0.27+
-    #Resaving seems to correct this as Klayout 26 will force a topcell (?)
+# def klayout_formatfile(input="output.gds"):
+#     import klayout
+#     #Due to a format update in Klayout 0.27+, newer version of Klayout cannot read the file format exported
+#     #by the lumerical encrypted library. The work around is to load the file in Klayout26 and resave it.
+#     #I believe the issue is because the Lumerical encrpyted library is saving multiple topcells
+#     #This can be seen from the errors when loading them directly into Klayout 0.27+
+#     #Resaving seems to correct this as Klayout 26 will force a topcell (?)
     
-    ly = db.Layout()
-    ly.read(input)
-    ly.write(input)
+#     ly = db.Layout()
+#     ly.read(input)
+#     ly.write(input)
 
 def klayout_mergefiles(directory="output",outputname = "lumexport.gds",removetempfiles=True):
     import os
-    ly = db.Layout()
-    top_cell = ly.create_cell("TOP")
+    try:
+        import klayout.db as db
+        installed = True
+    except ImportError:
+        os.system('cls' if os.name == 'nt' else 'clear')
+        print("[ERROR]: Klayout module not installed. Please refer to the 'Installation Instructions' in the readme.")
+        installed = False
     
-    for file in os.listdir(directory):
-        if file.endswith(".gds") and file.startswith("TEMP_output"):
-            ly_import = db.Layout()
-            ly_import.read(os.path.join(directory, file))
-            imported_top_cell = ly_import.top_cell()
+    if installed:
+        ly = db.Layout()
+        top_cell = ly.create_cell("TOP")
+        
+        for file in os.listdir(directory):
+            if file.endswith(".gds") and file.startswith("TEMP_output"):
+                ly_import = db.Layout()
+                ly_import.read(os.path.join(directory, file))
+                imported_top_cell = ly_import.top_cell()
 
-            target_cell = ly.create_cell(imported_top_cell.name)
-            target_cell.copy_tree(imported_top_cell)
+                target_cell = ly.create_cell(imported_top_cell.name)
+                target_cell.copy_tree(imported_top_cell)
 
-            # frees the resources of the imported layout
-            ly_import._destroy()
+                # frees the resources of the imported layout
+                ly_import._destroy()
 
-            inst = db.DCellInstArray(target_cell.cell_index(), db.DTrans(db.DVector(0, 0)))
-            top_cell.insert(inst)
-            
-            if removetempfiles:
-                os.remove(os.path.join(directory, file))
+                inst = db.DCellInstArray(target_cell.cell_index(), db.DTrans(db.DVector(0, 0)))
+                top_cell.insert(inst)
+                
+                if removetempfiles:
+                    os.remove(os.path.join(directory, file))
 
-    outputfullname = os.path.join(directory, outputname)
-    ly.write(outputfullname)
-    print("GDS files merged.")
-    return True       
+        outputfullname = os.path.join(directory, outputname)
+        ly.write(outputfullname)
+        print("GDS files merged.")
+
+        return True       
  
-if __name__ == '__main__':
+if __name__ == '__main__':     
+    print("This window opens because Lumerical's python command runs scripts through the Python Interpreter.\n")
+    
     if klayout_mergefiles():
         print("You may close this CMD window.")
+
     else:
-        print("Unexpected merging error. You may close this CMD window.")
+        print("[ERROR]: Unexpected error. You may close this CMD window.")
